Remove unnecessary X11 dependency
Since the snap requires mir-kiosk to run we know we don't need X11 support: imv is built for Wayland only. If I am not mistaken libxkbcommon-x11-dev is an X11 dependency that has been added in the past but is no longer needed.

I tried a clean build without libxkbcommon-x11-dev and everything still seems to work fine.